requirements.txt
torch
torchvision
numpy
pandas
opencv-python
requests
Pillow
transformers
timm
